check if svn_authz exists before potentially deleting a non-existent file

+++ lib/tasks/repos.rake
@@ -3,7 +3,9 @@ namespace :repos do
   task(:drop => :environment) do
     puts "Destroying Repositories..."
     FileUtils.rm_r Dir.glob(File.join(REPOSITORY_STORAGE, "*"))
-    File.delete($REPOSITORY_SVN_AUTHZ_FILE)
+    if File.exist?($REPOSITORY_SVN_AUTHZ_FILE)
+      File.delete($REPOSITORY_SVN_AUTHZ_FILE)
+    end
   end
   
   desc "Build repositories in REPOSITORY_STORAGE for all existing Groups"
